+++ src/Handler/EasyHandle.php
@@ -28,14 +28,14 @@ final class EasyHandle
     public $sink;
 
     /**
-     * @var array Received HTTP headers so far
+     * @var list<string> Received HTTP headers so far
      */
-    public $headers = [];
+    public array $headers = [];
 
     /**
      * @var ResponseInterface|null Received response (if any)
      */
-    public $response;
+    public ?ResponseInterface $response = null;
 
     /**
      * @var RequestInterface Request being sent
@@ -50,27 +50,27 @@ final class EasyHandle
     /**
      * @var int cURL error number (if any)
      */
-    public $errno = 0;
+    public int $errno = 0;
 
     /**
      * @var \Throwable|null Exception during on_headers (if any)
      */
-    public $onHeadersException;
+    public ?\Throwable $onHeadersException = null;
 
     /**
      * @var \Throwable|null Exception during progress callback (if any)
      */
-    public $progressException;
+    public ?\Throwable $progressException = null;
 
     /**
      * @var bool Whether the progress callback requested abort
      */
-    public $progressAborted = false;
+    public bool $progressAborted = false;
 
     /**
      * @var \Throwable|null Exception during createResponse (if any)
      */
-    public $createResponseException;
+    public ?\Throwable $createResponseException = null;
 
     /**
      * Attach a response to the easy handle based on the received headers.